perf: replace bit shift with rotation
The bit shift approach is inefficient for the following reasons: (1)
The shift amount given to a shift operator needs to be positive. This
means we need a conditional branch to handle negative shift amounts. (2)
In Rust, when shifting an `N`-bit integer, the actual shift amount is
determined by the lower `log2(N)` bits of the specified amount. However,
some target architectures work differently (i.e., all bits participate),
and therefore shifting on such architectures requires an extra
instruction to emulate the defined semantics.

This change brings about reduction in the code size as well as in the
execution time. In an experiment, it reduced the compiled size of
`SmallGlobalTlsf` by 52 bytes (on `wasm32`) and that of `Tlsf` by
12 bytes (on Cortex-M4) and the execution time of `Tlsf::allocate` +
`Tlsf::deallocate` by 1–2 cycles (on Cortex-M4).